cont1dof

DCMotor.cont1dof — Function

cont1dof(G, poles)

Calcula, mediante el método lineal algebraico, un controlador 1-GDL de realimentación unitaria que ubica los polos de lazo cerrado, según los polos indicados por el vector poles.

Argumentos

G: Función de transferencia de la planta.

poles::Vector: Polos deseados de lazo cerrado del sistema.

Retorna

C: función de transferencia del controlador.

Lanza

ErrorException si poles no tiene al menos 2n-1 elementos (con n el orden de G), es decir, si el orden del controlador resulta menor que n - 1.

Notas

Si selecciona un orden del controlador tal que m > n-1, el algoritmo coloca automaticamente acción integral

en el controlador diseñado.

Ejemplo

Primero, asegúrese de haber importado el paquete DCMotor, de haber definido el sistema y de haber obtenido un modelo nominal de la planta usando las funciones de identificación get_model_step o get_model_prbs. Así podrá invocar la función de transferencia del sistema con la función tf:

G_angle = tf(sys, :speed)

Luego, defina los polos deseados de lazo cerrado y diseñe el controlador:

polos_lc = [-10+10im, -10-10im, -80]
C = cont1dof(G, polos_lc) 
set_controller(sys, C; output=:angle)

# Aqui vemos la respuesta al escalon en lazo cerrado
T = feedback(C*G, 1)
result = step_closed(sys; r0 = 0, r1 =100,  t0 = 0.5, t1 =2); 
stepinfo(result,T)