ss

ControlSystemsBase.ss — Method

ss(sys::MotorSystem, output::Symbol=:angle)

Retorna la realización en espacio de estado del modelo nominal (sin retardo) de la plataforma DCMotor.

Argumentos

sys::MotorSystem: objeto que representa la plataforma.

output::Symbol=:angle: variable de salida del modelo. Según el valor de output, la realización retornada puede ser:

:angle: el modelo que retorna tiene salida de ángulo, dado por:

$\begin{bmatrix} \dot x_1 \\ \dot x_2 \end{bmatrix} = \begin{bmatrix} - a & 0 \\ 1 & 0 \end{bmatrix} \begin{bmatrix} x_1 \\ x_2 \end{bmatrix} + \begin{bmatrix} b \\ 0 \end{bmatrix} u$

$y = \begin{bmatrix} 0 & 1 \end{bmatrix}\begin{bmatrix} x_1 \\ x_2 \end{bmatrix}$.

:speed: el modelo que retorna es de velocidad angular, dado por:

$\dot x_1 = - a\,x_1 + b\,u$

$y = x_1$

Retorna

sys_ss::ControlSystemsBase.StateSpace: realización continua en espacio de estado de la plataforma DCMotor para la salida seleccionada.

Nota

Los parámetros $a$ y $b$ usados en las realizaciones se toman de los valores estimados por medio de las funciones get_model_step o get_model_prbs.

Ejemplos

sys = MotorSystem();
ss_angle = ss(sys, :angle)     # modelo del ángulo
ss_speed = ss(sys, :speed)     # modelo en velocidad angular