volts_from_speed

DCMotor.volts_from_speed — Function

volts_from_speed(sys::MotorSystem, speed::Real)

Estima el voltaje constante necesario para que el motor DC alcance, en estado estacionario, la velocidad angular indicada en la variable speed (en °/s). Es la función inversa de speed_from_volts y usa el mismo modelo estático lineal a trozos dado por:

$\qquad u=\dfrac{\omega_{ee}- \text{sign}(\omega_{ee})\,b}{K}$

Argumentos

sys::MotorSystem: objeto que representa la plataforma

speed::Real: velocidad angular de estado estacionario especificada (en °/s). Debe ser inferior a la máxima velocidad medida experimentalmente (típicamente cercana a 800 °/s a 5V), es decir,

|speed| $\leq 800 °/s$

Retorna

volts::Float64: voltaje [V] estimado requerido para alcanzar la velocidad angular de estado estacionario definida en la variable speed,.

Lanza

ErrorException si |speed| excede el rango medido experimentalmente.

Ejemplos

using DCMotor
sys = MotorSystem();
volts_from_speed(sys, 400.0)

    2.87